@ggerganov 2023-12-01
Looking for Apple Silicon runs of the following 7B benches, `master` vs `PR`:

```bash
make -j batched-bench && ./batched-bench ./models/llama-7b-v2/ggml-model-q8_0.gguf 10880 1 99 0 128,2048 128 1,2,4,8
```

```bash
make -j batched-bench && ./batched-bench ./models/llama-7b-v2/ggml-model-q4_0.gguf 10880 1 99 0 128,2048 128 1,2,4,8
```

cc @slaren @jhen0409 and everyone else with a Mac