@jhen0409 2023-12-02
M1 Max MBP (32c):

```bash
make -j batched-bench && ./batched-bench ./models/llama-7b-v2/ggml-model-q8_0.gguf 10880 1 99 0 128,2048 128 1,2,4,8
```

`master`
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |    128 |    1 |    256 |    0.247 |   517.85 |    3.205 |    39.94 |    3.452 |    74.16 |
|   128 |    128 |    2 |    384 |    0.248 |   516.60 |   10.249 |    24.98 |   10.497 |    36.58 |
|   128 |    128 |    4 |    640 |    0.246 |   519.61 |   10.387 |    49.29 |   10.633 |    60.19 |
|   128 |    128 |    8 |   1152 |    0.246 |   520.45 |   10.785 |    94.94 |   11.031 |   104.43 |
|  2048 |    128 |    1 |   2176 |    3.901 |   525.00 |    3.707 |    34.53 |    7.608 |   286.01 |
|  2048 |    128 |    2 |   2304 |    3.895 |   525.76 |   11.045 |    23.18 |   14.940 |   154.22 |
|  2048 |    128 |    4 |   2560 |    3.891 |   526.37 |   11.457 |    44.69 |   15.348 |   166.80 |
|  2048 |    128 |    8 |   3072 |    3.889 |   526.55 |   12.349 |    82.92 |   16.238 |   189.18 |

`PR`
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |    128 |    1 |    256 |    0.248 |   515.87 |    3.191 |    40.11 |    3.439 |    74.43 |
|   128 |    128 |    2 |    384 |    0.248 |   516.49 |    5.758 |    44.46 |    6.006 |    63.94 |
|   128 |    128 |    4 |    640 |    0.247 |   518.64 |   10.320 |    49.61 |   10.567 |    60.57 |
|   128 |    128 |    8 |   1152 |    0.246 |   519.55 |   19.319 |    53.00 |   19.565 |    58.88 |
|  2048 |    128 |    1 |   2176 |    3.906 |   524.27 |    3.695 |    34.64 |    7.602 |   286.25 |
|  2048 |    128 |    2 |   2304 |    3.898 |   525.39 |    6.510 |    39.32 |   10.408 |   221.36 |
|  2048 |    128 |    4 |   2560 |    3.891 |   526.30 |   11.049 |    46.34 |   14.940 |   171.35 |
|  2048 |    128 |    8 |   3072 |    3.891 |   526.30 |   20.171 |    50.77 |   24.062 |   127.67 |

```bash
make -j batched-bench && ./batched-bench ./models/llama-7b-v2/ggml-model-q4_0.gguf 10880 1 99 0 128,2048 128 1,2,4,8
```

`master`
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |    128 |    1 |    256 |    0.249 |   513.28 |    2.092 |    61.19 |    2.341 |   109.35 |
|   128 |    128 |    2 |    384 |    0.249 |   514.12 |   10.313 |    24.82 |   10.562 |    36.36 |
|   128 |    128 |    4 |    640 |    0.248 |   516.87 |   10.473 |    48.89 |   10.720 |    59.70 |
|   128 |    128 |    8 |   1152 |    0.247 |   518.07 |   10.871 |    94.20 |   11.118 |   103.62 |
|  2048 |    128 |    1 |   2176 |    3.978 |   514.80 |    2.620 |    48.85 |    6.599 |   329.76 |
|  2048 |    128 |    2 |   2304 |    3.972 |   515.62 |   11.055 |    23.16 |   15.027 |   153.33 |
|  2048 |    128 |    4 |   2560 |    3.968 |   516.12 |   11.518 |    44.45 |   15.486 |   165.31 |
|  2048 |    128 |    8 |   3072 |    3.967 |   516.27 |   12.458 |    82.20 |   16.425 |   187.04 |

`PR`
|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   128 |    128 |    1 |    256 |    0.255 |   501.64 |    2.124 |    60.27 |    2.379 |   107.61 |
|   128 |    128 |    2 |    384 |    0.249 |   515.07 |    3.555 |    72.01 |    3.803 |   100.96 |
|   128 |    128 |    4 |    640 |    0.248 |   515.99 |    6.051 |    84.61 |    6.299 |   101.60 |
|   128 |    128 |    8 |   1152 |    0.247 |   517.46 |   11.131 |    92.00 |   11.378 |   101.25 |
|  2048 |    128 |    1 |   2176 |    3.982 |   514.36 |    2.620 |    48.85 |    6.602 |   329.60 |
|  2048 |    128 |    2 |   2304 |    3.979 |   514.73 |    4.363 |    58.67 |    8.342 |   276.19 |
|  2048 |    128 |    4 |   2560 |    3.974 |   515.33 |    6.885 |    74.36 |   10.859 |   235.74 |
|  2048 |    128 |    8 |   3072 |    3.975 |   515.21 |   11.925 |    85.87 |   15.900 |   193.21 |


q8_0 drops significantly at B=8.